==> analysis/step0_preprocess_lookit_json.Rmd <==
---
title: "Preprocess lookit data from GARDEN vocab study"
author: "Bria Long"
date: "2025-03-11"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(jsonlite)
library(tidyverse)
library(here)
```

THIS USES IDENTIFIABLE DATA AND IS NOT SHAREABLE.
```{r}
lookit_json = here::here('data/lookit/Visual-Vocabulary--A-picture-matching-game-_all-responses-identifiable_072924_final.json')
responses = jsonlite::read_json(lookit_json)
```

Get back participant data from the json.
```{r}
participants = sequence(c(length(responses)))

for (c in participants){

start_visual_vocab = responses[[c]]$exp_data$`5-instructions-detailed`$eventTimings[[1]]$timestamp

end_visual_vocab = responses[[c]]$exp_data$`6-visualvocab`$eventTimings[[1]]$timestamp

eligible = responses[[c]]$response$eligibility[1]
birthdate_diff = responses[[c]]$response$birthdate_difference[1]

consent = responses[[c]]$consent$ruling
child_hashed_id = responses[[c]]$child$hashed_id
child_name = responses[[c]]$child$name
child_global_id = responses[[c]]$child$global_id
age_in_days = responses[[c]]$child$age_in_days
if (is.null(child_name)==TRUE) {child_name = NA}
parent_name = responses[[c]]$participant$nickname
parent_global_id = responses[[c]]$participant$global_id
parent_hashed_id = responses[[c]]$participant$hashed_id
completed = responses[[c]]$response$completed
feedback = responses[[c]]$response$parent_feedback
session_id = responses[[c]]$response$uuid

this_child = tibble(response = c, start_timestamp = start_visual_vocab, end_timestamp = end_visual_vocab, child_id = child_hashed_id, parent_id = parent_hashed_id, completed = completed, child_name = child_name, parent_name = parent_name, consent = consent, eligible = eligible, birthdate_difference = birthdate_diff, parent_feedback = feedback, child_global_id = child_global_id, parent_global_id = parent_global_id, age_in_days = age_in_days, session_id = session_id)

if (c==1) {all_children = this_child}

if (is.null(start_visual_vocab)==FALSE && is.null(end_visual_vocab)==FALSE){
all_children = all_children %>%
  full_join(this_child)
}

}
```
# for loading back for compensation
```{r}
save(all_children, file = paste0(here::here('data/preprocessed/all_children_identifiable'), today(), '.RData'))
```
